Commit Day 9

+++ AdventOfCode/Base/IPuzzleInput.cs
@@ -13,6 +13,6 @@ namespace AoC.AdventOfCode.Base
         void LoadPuzzleInput(string input);
         void LoadPuzzleInputFromFile(string input);
         void LoadAdditionalParameter(string[] args);
-        int SolvePuzzle(int part = 0);
+        long SolvePuzzle(int part = 0);
     }
 }
